DOI / Add API to check status and create DOI.

+++ doi/src/main/java/org/fao/geonet/doi/client/DoiManager.java
@@ -55,6 +55,7 @@ import java.util.Map;
  */
 public class DoiManager {
     private static final String DOI_ADD_XSL_PROCESS = "process/doi-add.xsl";
+    private static final String DOI_REMOVE_XSL_PROCESS = "process/doi-remove.xsl";
     public static final String DATACITE_XSL_CONVERSION_FILE = "formatter/datacite/view.xsl";
     public static final String DOI_PREFIX_PARAMETER = "doiPrefix";
     public static final String HTTPS_DOI_ORG = "https://doi.org/";
@@ -132,18 +133,44 @@ public class DoiManager {
     }
 
 
-    public void register(ServiceContext context, AbstractMetadata metadata) throws Exception {
+    public Map<String, Boolean> check(ServiceContext serviceContext, AbstractMetadata metadata, Element dataciteMetadata) throws Exception {
+        Map<String, Boolean> conditions = new HashMap<>();
+        checkInitialised();
+        conditions.put(DoiConditions.API_CONFIGURED, true);
+
+        String doi =  DoiBuilder.create(this.doiPrefix, metadata.getUuid());
+        checkPreConditions(metadata, doi);
+        conditions.put(DoiConditions.RECORD_IS_PUBLIC, true);
+        conditions.put(DoiConditions.STANDARD_SUPPORT, true);
+
+
+        // ** Convert to DataCite format
+        Element dataciteFormatMetadata =
+            dataciteMetadata == null ?
+            convertXmlToDataCiteFormat(metadata.getDataInfo().getSchemaId(),
+                metadata.getXmlData(false)) : dataciteMetadata;
+        checkPreConditionsOnDataCite(metadata, doi, dataciteFormatMetadata);
+        conditions.put(DoiConditions.DATACITE_FORMAT_IS_VALID, true);
+        return conditions;
+    }
+
+    public Map<String, String> register(ServiceContext context, AbstractMetadata metadata) throws Exception {
+        Map<String, String> doiInfo = new HashMap<>(3);
         // The new DOI for this record
         String doi =  DoiBuilder.create(this.doiPrefix, metadata.getUuid());
+        doiInfo.put("doi", doi);
 
-        checkInitialised(); // DOI Configuration ok ?
+        // The record in datacite format
+        Element dataciteFormatMetadata =
+                convertXmlToDataCiteFormat(metadata.getDataInfo().getSchemaId(),
+                    metadata.getXmlData(false));
 
-        checkPreConditions(metadata, doi);
-        checkPreConditionsOnDataCite(metadata, doi);
 
-        createDoi(context, metadata, doi);
+        check(context, metadata, dataciteFormatMetadata);
+        createDoi(context, metadata, doiInfo, dataciteFormatMetadata);
+        checkDoiCreation(metadata, doiInfo);
 
-        checkDoiCreation(metadata, doi);
+        return doiInfo;
     }
 
     /**
@@ -163,13 +190,14 @@ public class DoiManager {
             visibleToAll = am.isVisibleToAll(metadata.getId() + "");
         } catch (Exception e) {
             throw new DoiClientException(String.format(
-                "Failed to check if record '%s' is visible to all for DOI creation. Error is %s.",
+                "Failed to check if record '%s' is visible to all for DOI creation." +
+                   " Error is %s.",
                 metadata.getUuid(), e.getMessage()));
         }
 
         if (!visibleToAll) {
             throw new DoiClientException(String.format(
-                "Record '%s' is not public and we cannot request a DOI for such a record.",
+                "Record '%s' is not public and we cannot request a DOI for such a record. Publish this record first.",
                 metadata.getUuid()));
         }
 
@@ -183,12 +211,18 @@ public class DoiManager {
                 // Current doi does not match the one going to be inserted. This is odd
                 if (!currentDoi.equals(doi)) {
                     throw new DoiClientException(String.format(
-                        "Record '%s' already contains a DOI '%' which is not equal to the DOI about to be created. Maybe current DOI does not correspond to that record? This may happen when creating a copy of a record having an existing DOI.",
+                        "Record '%s' already contains a DOI '%' which is not equal " +
+                            "to the DOI about to be created (ie. '%s'). " +
+                            "Maybe current DOI does not correspond to that record? " +
+                            "This may happen when creating a copy of a record having " +
+                            "an existing DOI.",
                         metadata.getUuid(), currentDoi, doi));
                 }
 
                 throw new DoiClientException(String.format(
-                    "Record '%s' already contains a DOI. The DOI is '%s'. We cannot register it again.",
+                    "Record '%s' already contains a DOI. The DOI is '%s'. " +
+                        "We cannot register it again. " +
+                        "Remove the DOI reference if it does not apply to that record.",
                     metadata.getUuid(), currentDoi));
             }
         } catch (ResourceNotFoundException e) {
@@ -200,7 +234,9 @@ public class DoiManager {
             //                value="*//gmd:CI_OnlineResource[gmd:protocol/gco:CharacterString = 'WWW:DOI']/gmd:URL/text()"/>
             //            </bean>
             throw new DoiClientException(String.format(
-                "Record '%s' is in schema '%s' and we cannot find a saved query with id '%s' to retrieve the DOI. Error is %s. Check the schema %sSchemaPlugin and add the DOI get query.",
+                "Record '%s' is in schema '%s' and we cannot find a saved query " +
+                    "with id '%s' to retrieve the DOI. Error is %s. " +
+                    "Check the schema %sSchemaPlugin and add the DOI get query.",
                 metadata.getUuid(), schema.getName(),
                 DOI_GET_SAVED_QUERY, e.getMessage(),
                 schema.getName()));
@@ -212,13 +248,34 @@ public class DoiManager {
      * Check conditions on DataCite side.
      * @param metadata
      * @param doi
+     * @param dataciteMetadata
      */
-    private void checkPreConditionsOnDataCite(AbstractMetadata metadata, String doi) {
+    private void checkPreConditionsOnDataCite(AbstractMetadata metadata, String doi, Element dataciteMetadata) throws DoiClientException {
         // * DataCite API is up an running ?
+
+        // XSD validation
+        try {
+            Xml.validate(dataciteMetadata);
+        } catch (Exception e) {
+            throw new DoiClientException(String.format(
+                "Record '%s' converted to DataCite format is invalid. Error is: %s. " +
+                    "Required fields in DataCite are: identifier, creators, titles, publisher, publicationYear, resourceType. Check the DataCite format output at " +
+                    "%sapi/records/%s/formatters/datacite?output=xml and " +
+                    "adpat the record content to add missing information.",
+                metadata.getUuid(), e.getMessage(), sm.getNodeURL(), metadata.getUuid()));
+        }
 
         // * MDS / DOI does not exist already
         // curl -i --user username:password https://mds.test.datacite.org/doi/10.5072/GN
         // Return 404
+        final String doiResponse = client.retrieveDoi(doi);
+        if (doiResponse != null) {
+            throw new DoiClientException(String.format(
+                "Record '%s' looks to be already published on DataCite on DOI '%s'. Check DataCite DOI: %s. " +
+                    "If the DOI is not correct, remove it from the record and ask for a new one.",
+                metadata.getUuid(), doi, doiResponse));
+        }
+        // TODO: Could be relevant at some point to return states (draft/findable)
 
         // * MDS / Metadata does not exist either
         // curl -i --user username:password https://mds.test.datacite.org/metadata/10.5072/GN
@@ -229,32 +286,30 @@ public class DoiManager {
      * Use the DataCite API to register the new DOI.
      * @param context
      * @param metadata
-     * @param doi
      */
-    private void createDoi(ServiceContext context, AbstractMetadata metadata, String doi) throws Exception {
+    private void createDoi(ServiceContext context, AbstractMetadata metadata, Map<String, String> doiInfo, Element dataciteMetadata) throws Exception {
         // * Now, let's create the DOI
         // picking a DOI name,
 
         // The identifier in the DataCite XML metadata file can be left empty, as the value from the PUT URL will be used.
         // curl -H "Content-Type: application/xml;charset=UTF-8" -X POST -i --user username:password -d @/tmp/GN.xml https://mds.test.datacite.org/metadata/10.5072/GN
 
-        // ** Convert to DataCite format
-        Element dataciteMetadata =
-            convertXmlToDataCiteFormat(metadata.getDataInfo().getSchemaId(),
-            metadata.getXmlData(false));
         // ** Validate output ? XSD
         // ** POST metadata
         // 201 Created: operation successful,
-        client.createDoiMetadata(doi, Xml.getString(dataciteMetadata));
+        client.createDoiMetadata(doiInfo.get("doi"), Xml.getString(dataciteMetadata));
 
 
         // Register the URL
         // 201 Created: operation successful;
-        client.createDoi(doi, landingPageTemplate.replace("{{uuid}}", metadata.getUuid()));
+        String landingPage = landingPageTemplate.replace(
+                        "{{uuid}}", metadata.getUuid());
+        doiInfo.put("doiLandingPage", landingPage);
+        client.createDoi(doiInfo.get("doi"), landingPage);
 
 
         // Add the DOI in the record
-        Element recordWithDoi = setDOIValue(doi, metadata.getDataInfo().getSchemaId(), metadata.getXmlData(false));
+        Element recordWithDoi = setDOIValue(doiInfo.get("doi"), metadata.getDataInfo().getSchemaId(), metadata.getXmlData(false));
         // Update the published copy
         //--- needed to detach md from the document
 //        md.detach();
@@ -270,32 +325,38 @@ public class DoiManager {
      * @param metadata
      * @param doi
      */
-    private void checkDoiCreation(AbstractMetadata metadata, String doi) {
+    private void checkDoiCreation(AbstractMetadata metadata, Map<String, String> doi) {
         // Check it is available on DataCite Metadata Store
         // curl -X GET --user INIST.IFREMER https://mds.test.datacite.org/metadata/10.5072/GN2
         // Check it is in the record
     }
 
 
-    /**
-     * Unregisters a DOI.
-     *
-     * @param doi
-     */
-    private void unregisterDoi(String doi) throws DoiClientException {
+    public void unregisterDoi(AbstractMetadata metadata, ServiceContext context) throws DoiClientException, ResourceNotFoundException {
         checkInitialised();
 
-        client.deleteDoiMetadata(doi);
-    }
+        final String doi = DoiBuilder.create(this.doiPrefix, metadata.getUuid());
+        final String doiResponse = client.retrieveDoi(doi);
+        if (doiResponse == null) {
+            throw new ResourceNotFoundException(String.format(
+                "Record '%s' is not available on DataCite. DOI '%s' does not exist.",
+                metadata.getUuid(), doi));
+        }
 
-    public void unregisterDoi(Metadata metadata, Element md) throws DoiClientException {
-        checkInitialised();
 
         try {
-            String doi = schemaUtils.getSchema(metadata.getDataInfo().getSchemaId())
+            Element md = metadata.getXmlData(false);
+
+            String doiUrl = schemaUtils.getSchema(metadata.getDataInfo().getSchemaId())
                 .queryString(DOI_GET_SAVED_QUERY, md);
-            unregisterDoi(doi);
-            // TODO: Remove from record ?
+
+            client.deleteDoiMetadata(doi);
+            client.deleteDoi(doi);
+
+            Element recordWithoutDoi = removeDOIValue(doiUrl, metadata.getDataInfo().getSchemaId(), md);
+
+            dm.updateMetadata(context, metadata.getId() + "", recordWithoutDoi, false, true, true,
+                context.getLanguage(), new ISODate().toString(), true);
         } catch (Exception ex) {
             throw new DoiClientException(ex.getMessage());
         }
@@ -318,6 +379,22 @@ public class DoiManager {
         return Xml.transform(md, styleSheet, params);
     }
 
+    /**
+     * Sets the DOI URL value in the metadata record using the process DOI_ADD_XSL_PROCESS.
+     *
+     */
+    public Element removeDOIValue(String doi, String schema, Element md) throws Exception {
+        Path styleSheet = dm.getSchemaDir(schema).resolve(DOI_REMOVE_XSL_PROCESS);
+        boolean exists = Files.exists(styleSheet);
+        if (!exists) {
+            throw new DoiClientException(String.format("To remove a DOI, the schema has to defined how to remove a DOI in the record. The schema_plugins/%s/process/%s was not found. Create the XSL transformation.",
+                schema, DOI_REMOVE_XSL_PROCESS));
+        }
+
+        Map<String, Object> params = new HashMap<>(1);
+        params.put("doi", doi);
+        return Xml.transform(md, styleSheet, params);
+    }
 
     /**
      * Convert a metadata record to the DataCite Metadata Schema.
@@ -342,7 +419,8 @@ public class DoiManager {
 
     private void checkInitialised() throws DoiClientException {
         if (!initialised) {
-            throw new DoiClientException("DOI configuration is not complete. Check in System Configuration to fill the DOI configuration.");
+            throw new DoiClientException("DOI configuration is not complete. Check System Configuration and set the DOI configuration.");
         }
     }
+
 }
